Add createIsStaffSelector to authSelectors
Makes working with staff permissions a lot easier and
straightforward.
Refs #506.
