+++ app/pages/user-reservations/reservation-list/reservationListSelector.js
@@ -1,11 +1,10 @@
 import { createStructuredSelector } from 'reselect';
 
 import ActionTypes from 'constants/ActionTypes';
-import { isAdminSelector } from 'state/selectors/authSelectors';
+import { isAdminSelector, staffUnitsSelector } from 'state/selectors/authSelectors';
 import { resourcesSelector, unitsSelector } from 'state/selectors/dataSelectors';
 import requestIsActiveSelectorFactory from 'state/selectors/factories/requestIsActiveSelectorFactory';
 import sortedReservationsSelector from 'state/selectors/sortedReservationsSelector';
-import staffUnitsSelector from 'state/selectors/staffUnitsSelector';
 
 const reservationListSelector = createStructuredSelector({
   isAdmin: isAdminSelector,
