+++ app/shared/reservation-confirmation/reservationConfirmationSelector.js
@@ -3,13 +3,14 @@ import { createSelector, createStructuredSelector } from 'reselect';
 
 import ActionTypes from 'constants/ActionTypes';
 import ModalTypes from 'constants/ModalTypes';
+import { createIsStaffSelector } from 'state/selectors/authSelectors';
 import { createResourceSelector } from 'state/selectors/dataSelectors';
 import selectedReservationsFromStateSelector from 'state/selectors/selectedReservationsSelector';
-import staffUnitsSelector from 'state/selectors/staffUnitsSelector';
 import modalIsOpenSelectorFactory from 'state/selectors/factories/modalIsOpenSelectorFactory';
 import requestIsActiveSelectorFactory from 'state/selectors/factories/requestIsActiveSelectorFactory';
 
 const resourceIdSelector = (state, props) => props.params.id;
+const resourceSelector = createResourceSelector(resourceIdSelector);
 const toEditSelector = state => state.ui.reservations.toEdit;
 const staffEventSelectedSelector = state => (
   formValueSelector('preliminaryReservation')(state, 'staffEvent')
@@ -23,11 +24,11 @@ const selectedReservationsSelector = createSelector(
 const reservationCalendarSelector = createStructuredSelector({
   confirmReservationModalIsOpen: modalIsOpenSelectorFactory(ModalTypes.RESERVATION_CONFIRM),
   isMakingReservations: requestIsActiveSelectorFactory(ActionTypes.API.RESERVATION_POST_REQUEST),
+  isStaff: createIsStaffSelector(resourceSelector),
   reservationsToEdit: toEditSelector,
-  resource: createResourceSelector(resourceIdSelector),
+  resource: resourceSelector,
   selectedReservations: selectedReservationsSelector,
   staffEventSelected: staffEventSelectedSelector,
-  staffUnits: staffUnitsSelector,
 });
 
 export default reservationCalendarSelector;
